SQLite fix (#3)
* Fix SQLite examples

* Move to SQLite Folder

add more packages

more libs

Revert "more libs"

This reverts commit 3aafc329a0f4cb1be243e38ae3db7bb5b28d3421.

* Fix runtime crash

+++ Sqlite/MauiSqlite/MainPage.xaml.cs
@@ -34,6 +34,9 @@ public partial class MainPage : ContentPage
             return;
 
         var account = collectionView.SelectedItem as Account;
+        if (account is null)
+            return;
+
         account.Balance = 0;
         _accountRepository.UpdateAccount(account);
         GetAccounts();
@@ -45,6 +48,9 @@ public partial class MainPage : ContentPage
             return;
 
         var account = collectionView.SelectedItem as Account;
+        if (account is null)
+            return;
+
         _accountRepository.DeleteAccount(account);
         GetAccounts();
     }
